change duration to long
git-svn-id: https://svn.win.tue.nl/repos/prom/Packages/InductiveMiner/Trunk@43794 a68f48d4-8d00-424f-b812-3f22a534a02a

+++ src/org/processmining/plugins/inductiveminer2/attributes/AttributeUtils.java
@@ -13,7 +13,7 @@ public class AttributeUtils {
 
 	public static long valueLong(Attribute attribute, XAttributable x) {
 		if (attribute.isDuration()) {
-			return (long) attribute.getDuration(x);
+			return attribute.getDuration(x);
 		} else if (attribute.isNumeric()) {
 			return (long) attribute.getNumeric(x);
 		} else if (attribute.isTime()) {
